Moves video and video index into seperate folders
