+++ src/App.js
@@ -1,19 +1,14 @@
 import Nav from './components/Nav/Nav';
 import Home from './components/Home/Home';
 import About from './components/About/About';
-import Video from './components/Video';
-import VideoIndex from './components/VideoIndex';
+import Video from './components/Video/Video';
+import VideoIndex from './components/VideoIndex/VideoIndex';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import { useState } from "react";
-import './index.css'
-
+import { useState } from 'react';
+import './index.css';
 
 export default function App() {
-
-  const [videos, setVideos ] = useState([])
-
-
-
+  const [videos, setVideos] = useState([]);
 
   return (
     <div className="wrapper">
