--- src/components/VideoIndex.css
@@ -1,34 +1,0 @@
-
-.thumbnailsDiv {
-    display: grid;
-    justify-items: center;
-    grid-template-columns: 480px 480px;
-    column-gap: 50px;
-    margin: 50px;
-}
-
-.thumbnail {
-    height: 480px;
-    display: flex;
-    flex-direction: column;
-    gap: 30px;
-}
-
-.thumbnailImg {
-    height: 360px;
-    width: 480px;
-}
-
-.title {
-    font-size: 20px;
-}
-
-img {
-    border-radius: 20px;
-    cursor: pointer;
-    transition: all ease 600ms;
-}
-
-img:hover {
-    border-radius: 0px;
-}
